Update ITokenService + impl ITokenService + update appsettings
