+++ CashaWeb/Casha.BLL/Casha.BLL.csproj
@@ -7,10 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="Services\" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\Casha.Core\Casha.Core.csproj" />
     <ProjectReference Include="..\Casha.DAL\Casha.DAL.csproj" />
   </ItemGroup>
