+++ CashaWeb/Casha.BLL/Interfaces/ITokenService.cs
@@ -1,10 +1,9 @@
 ﻿using Casha.Core.DbModels;
-using Microsoft.IdentityModel.Tokens;
 
 namespace Casha.BLL.Interfaces
 {
     public interface ITokenService
     {
-        public SecurityTokenDescriptor GenerateToken(User user);
+        public string GenerateToken(User user);
     }
 }
