+++ CashaWeb/Casha.BLL/Services/TokenService.cs
@@ -1,0 +1,51 @@
+﻿using Casha.BLL.Interfaces;
+using Casha.Core.DbModels;
+using Microsoft.Extensions.Configuration;
+using Microsoft.IdentityModel.Tokens;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
+
+namespace Casha.BLL.Services
+{
+    public class TokenService : ITokenService
+    {
+        private IConfiguration _config;
+
+        public TokenService(IConfiguration config)
+        {
+            _config = config;
+        }
+
+        public string GenerateToken(User user)
+        {
+            string issuer = _config["Jwt:Issuer"];
+            string audience = _config["Jwt:Audience"];
+            byte[] key = Encoding.ASCII.GetBytes(_config["Jwt:Key"]);
+
+            var claims = new List<Claim>
+                {
+                    new Claim("Id", user.Id),
+                    new Claim(JwtRegisteredClaimNames.Sub, user.UserName),
+                    new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
+                    new Claim(ClaimTypes.Role, "Admin")
+                };
+
+            var tokenDescriptor = new SecurityTokenDescriptor
+            {
+                Subject = new ClaimsIdentity(claims),
+                Expires = DateTime.UtcNow.AddDays(int.Parse(_config["Jwt:Duration"])),
+                Issuer = issuer,
+                Audience = audience,
+                SigningCredentials = new SigningCredentials
+                (new SymmetricSecurityKey(key),
+                    SecurityAlgorithms.HmacSha512Signature)
+            };
+
+            var tokenHandler = new JwtSecurityTokenHandler();
+            var token = tokenHandler.CreateToken(tokenDescriptor);
+            
+            return tokenHandler.WriteToken(token);
+        }
+    }
+}
