+++ CashaWeb/CashaWeb/Controllers/JwtController.cs
@@ -11,51 +11,12 @@ namespace CashaWeb.Controllers
     [ApiController]
     public class JwtController : Controller
     {
-        private IConfiguration _config;
-
-        public JwtController(IConfiguration config)
-        {
-            _config = config;
-        }
+        
 
         [HttpPost("login")]
         public async Task<IActionResult> LoginUser(LoginUser user)
         {
-            string issuer = _config["Jwt:Issuer"];
-            string audience = _config["Jwt:Audience"];
-            byte[] key = Encoding.ASCII.GetBytes(_config["Jwt:Key"]);
-
-            SecurityTokenDescriptor tokenDescriptor;
-            try
-            {
-                var claims = new List<Claim>
-                {
-                    new Claim("Id", Guid.NewGuid().ToString()),
-                    new Claim(JwtRegisteredClaimNames.Sub, user.Login),
-                    new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
-                    new Claim(ClaimTypes.Role, "Admin")
-                };
-
-                tokenDescriptor = new SecurityTokenDescriptor
-                {
-                    Subject = new ClaimsIdentity(claims),
-                    Expires = DateTime.UtcNow.AddDays(1),
-                    Issuer = issuer,
-                    Audience = audience,
-                    SigningCredentials = new SigningCredentials
-                    (new SymmetricSecurityKey(key),
-                        SecurityAlgorithms.HmacSha512Signature)
-                };
-            }
-            catch (NullReferenceException ex)
-            {
-                return BadRequest(ex.Message + " Emploee not found.");
-            }
-
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var token = tokenHandler.CreateToken(tokenDescriptor);
-            var jwtToken = tokenHandler.WriteToken(token);
-            var stringToken = tokenHandler.WriteToken(token);
+            
 
             return Ok(stringToken);
