+++ CashaWeb/CashaWeb/appsettings.json
@@ -11,7 +11,8 @@
   "Jwt": {
     "Issuer": "",
     "Audience": "",
-    "Key": "Long Unknown Secret Word"
+    "Key": "Long Unknown Secret Word",
+    "Duration": "1"
   }
 
   "AllowedHosts": "*"
